+++ lib/guis/static_gui/widgets/SetpointWidget.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "../lvgl_helpers.hpp"
+#include "../../lvgl_helpers.hpp"
 #include "./BaseWidget.hpp"
 #include "blocks/SetpointSensorPairBlock.hpp"
